+++ source/public/Backup-MssqlDatabase.ps1
@@ -161,5 +161,19 @@ function Backup-MssqlDatabase {
         Get-CurrentVariables -InitialVariables $initialVariables -Debug:$DebugPreference
         Return $result
     }
-    # Test:  Backup-MssqlDatabase -udl 'C:\temp\Eulanda_1 JohnDoe.udl' -storageFolder 'C:\store' -server 'mysftp.eulanda.eu' -user 'johndoe' -password 'superPass'
+
+    <#
+
+        $pesterFolder = Resolve-Path -path ".\source\tests"
+        $iniPath = Join-Path -path $pesterFolder "pester.ini"
+        $ini = Read-IniFile -path $iniPath
+        $path = $ini['SFTP']['SecurePasswordPath']
+        $path = $path -replace '\$home', $HOME
+        $secure = Import-Clixml -path $path
+        $server = $ini['SFTP']['Server']
+        $user = $ini['SFTP']['User']
+        Backup-MssqlDatabase -udl "$pesterFolder\Eulanda_1 Pester.udl" -storageFolder 'C:\store' -server $server -user $user -password $secure -remoteFolder '/inbox'
+
+    #>
+
 }