+++ source/public/Remove-RemoteFolder.ps1
@@ -58,19 +58,16 @@ Function Remove-RemoteFolder {
 
    <# Test:
 
-        $Features = Import-Module -Name '.\EulandaConnect.psm1' -PassThru -Force
-        & $Features {
-            $pesterFolder = Resolve-Path -path ".\source\tests"
-            $iniPath = Join-Path -path $pesterFolder "pester.ini"
-            $ini = Read-IniFile -path $iniPath
-            $path = $ini['SFTP']['SecurePasswordPath']
-            $path = $path -replace '\$home', $HOME
-            $secure = Import-Clixml -path $path
-            $server = $ini['SFTP']['Server']
-            $user = $ini['SFTP']['User']
+        $pesterFolder = Resolve-Path -path ".\source\tests"
+        $iniPath = Join-Path -path $pesterFolder "pester.ini"
+        $ini = Read-IniFile -path $iniPath
+        $path = $ini['SFTP']['SecurePasswordPath']
+        $path = $path -replace '\$home', $HOME
+        $secure = Import-Clixml -path $path
+        $server = $ini['SFTP']['Server']
+        $user = $ini['SFTP']['User']
 
-            Remove-RemoteFolder -server $server -protocol sftp -user $user -password $secure -remoteFolder /inbox/pester
-        }
+        Remove-RemoteFolder -server $server -protocol sftp -user $user -password $secure -remoteFolder /inbox/pester
 
     #>
 }